Move export and logger out of component folder
