+++ .prettierignore
@@ -1,0 +1,2 @@
+compass/
+node_modules/
